nsone/requirements.txt
ns1-python==0.9.19